python/whisper-api/pyproject.toml
[project]
name = "whisper-api"
version = "0.1.0"
description = "Simple FastAPI endpoint for OpenAI Whisper transcription"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "openai",
    "python-dotenv",
    "pydantic",
    "pydub",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
